fix: Refactor CI/CD workflows
We now have a single workflow that always runs Build and Analyze jobs.
If those succeed, and it's the main branch, run a deploy to prod.

Deployments no longer depend on release branch naming.
Version bumping workflow has been removed for now (bump manually).
For now, deployments are only triggered by pushes to main.

Follow-up later and add a dry-run deployment for PRs, as well as a dev
env deploy.

### Changes

- Add iceberg utils as git submodule
- Combine all workflows into a single "CICD" workflow
- Add makefile for linting, building, deploying
- Move all infra/function/ configs into root of the repo
- Update GH Actions workflow to Java 17 to match function code and runtime
- Enable auto publishing the "latest" alias for the function
- Add a few more CFN stack parameters and TODO notes
- Set MinValue and MaxValue for certain numerical paramters
- Disable SAM CLI telemetry in Makefile